+++ lib/bali/dsl/map_rules_dsl.rb
@@ -42,8 +42,12 @@ class Bali::MapRulesDsl
     raise Bali::DslError, "describe block must be within rules_for block"
   end
 
+  def role(*params)
+    raise Bali::DslError, "role block must be within rules_for block"
+  end
+
   def can(*params)
-    raise Bali::DslError, "can block must be within describe block"
+    raise Bali::DslError, "can block must be within role block"
   end
 
   def cant(*params)
@@ -52,15 +56,15 @@ class Bali::MapRulesDsl
   end
 
   def cannot(*params)
-    raise Bali::DslError, "cant block must be within describe block"
+    raise Bali::DslError, "cant block must be within role block"
   end
 
   def can_all(*params)
-    raise Bali::DslError, "can_all block must be within describe block"
+    raise Bali::DslError, "can_all block must be within role block"
   end
 
   def clear_rules
-    raise Bali::DslError, "clear_rules must be called within describe block"
+    raise Bali::DslError, "clear_rules must be called within role block"
   end
 
   def cant_all(*params)
@@ -69,6 +73,6 @@ class Bali::MapRulesDsl
   end
 
   def cannot_all(*params)
-    raise Bali::DslError, "cant_all block must be within describe block"
+    raise Bali::DslError, "cant_all block must be within role block"
   end
 end